Split out contributed custom editors values to own file
